feat(buildcop): add support for Python (#237)
Python has a slightly different flavor of xUnit output. So, I updated
the parsing to match (testsuites and package name handling).

I switched the XML parsing to be compact. It makes iterating over the
results a bit cleaner to not have to worry about 'elements'. There is no
loss of accuracy for this purpose. The not-compact version only seems
needed if you need stable parsing and re-printing.

The Python test results also come in many different log files. So, I
changed the logic for checking when a test passed to be an explicit pass
rather than just not-fail. A possible downside of this is that if a
failing test is renamed, the corresponding issue will not be
automatically closed/renamed.