fix(Build Cop): handle empty testsuites, better & more logging
Some logs have empty testsuites and it was causing errors.

Before, every invocation would comment about the same issues being flaky
and not closing them. By moving it below the pass/fail check, we'll only
log the message when the test explicitly passed.

I added the build URL to a few log messages to be able to look up the
corresponding build and raw XML output.

I also added a log message to the very beginning of the invocation to
make it easier to correlate particular builds with errors.