fix(merge-on-green): payload switches between Buffer and object
In debugging the issues with merge on green on GoogleCloudPlatform,
we ultimately were able to isolate the problem to the fact that
Cloud Scheduler tasks for googlapis were parsed as objects upon
arrival to our Cloud Function, whereas payloads for GoogleCloudPlatform
were still a Buffer.

We already have a helper that conditionally decodes the payload
appropriately let's use this.

Note: we should still try to root cause this, if only because it's
bizarre.