fix: upgrade header-checker-lint to probot 10 (#1006)
The one substantial change I made was to add an output to the check run that runs if the check is successful. The type change for that one was a bit annoying, and this was the simplest way I saw to fix it. Feel free to make changes as you see fit!