fix(merge-on-green): do not merge PRs with DNM label
This is a complement to the new Do Not Merge bot, which adds a failing
check to PRs with the DNM label. However, that does not block MoG from
merging the PR.

I chose to check for the label exactly, rather than checking for a
failing DNM check. If we end up with more than one relevant label (e.g.
do not submit), we'll have to "duplicate" that list. That seems better
than depending on the exact name of the DNM check.

Also, the DNM bot may not be installed everywhere MoG is, but the DNM
label has very obvious semantics MoG should follow.