feat: support multiple .OwlBot.yamls in one repo
Also deleted CopyTasksStore, because it was never used.  To avoid duplicate copy operations, the old code searched for source commit hashes, and the new code searchs for copy tags.

Replaced the yamls table (firestore collection) with a repo-configs table.
The schema changed enough so that the new code would be incompatible with the old schema.  We must run scan-configs once before releasing the new docker image, to populate the new table.