fix(owl-bot): bump the maxRetries to 30 (#2232)
We performed a real world stress tests with 80 PRs labeled with
`owlbot:run`. While owl-bot finished all the jobs, some tasks are
retried 4 or 5 times already.

Since we made the owl-bot check a mandatory check, we should minimize
the posibility of discarding the job. Maybe we should change to pull
model, but it needs some design changes.