fix(snippet-bot): skip errors when fetching renamed files
part of #2246

This is not a complete solution, but pragmatically it will work well.
The bot used to die when there is a file larger than 1MB.

With this PR, the bot ignores such an error. Most of the cases, those
large files are data files and they normally don't include region tags.