feat: flag tells owl-bot to squash first post-processor update
As requested by .NET maintainers.  They don't want to see all the
changes to the staging directory in their git history.

Here's how it works:
1. Owl Bot copies code from googleapis-gen and creates pull request in
   client library repo.  The pull request body contains a Copy-Tag.
2. Owl Bot runs the post processor as usual on the pull request.
3. Owl Bot examines the Copy-Tag in the most recent commit only.
   Most commits will not have a Copy-Tag, in which case jump to step 6.
4. Owl Bot inspects the .OwlBot.yaml file specified by the Copy-Tag.
5. If the .OwlBot.yaml has `squash: true`, then Owl Bot squashes the
   changes introduced by the post processor.
6. Otherwise, Owl Bot adds a new commit as usual.
7. Owl Bot pushes the changes back to the pull request.

To avoid a 3-minute window of brokenness, this pull request does not
contain the necessary changes to update-pr.yaml.
Another pull request will replace the last 4 commands in update-pr.yaml
with a call to commit-post-processor-update.