test: add container test for bazel-bot image (#2998)
This adds a very basic check that the Docker image we build for bazel-bot is valid. It verifies that `bazel`, `git`, `gh`, `jq`, and `jwt` binaries run `--version` and complete successfully. In the future, we could be more specific and check for exact versions.

Upon merge, our automatic cloud build trigger sync will create a presubmit trigger for the `packages/bazel-bot-docker-image` directory and run this test job.

Fixes #2995