fix(owl-bot): only use lock around running post-processor (#3316)
Currently, owlbot obtains a lock on a owner/repo/sha as soon as an event that could trigger a post-processor event runs. Instead, it should only attempt to run the post-processor under a lock.

This change moves the lock acquisition to a helper that wraps the run post-processor logic. This allows the webhook handlers to do their pre-checks and clean up no matter if it actually ends up running the post-processor or not.

Fixes #3257