fix: when logging errors, log all errors contained in an AggregateError (#3933)
Probot can run multiple handlers for a single event. It wraps any thrown errors in an `AggregateError`. When handling an `AggregateError`, log each contained error so Error Reporting will pick up each one individually.

This should be a huge quality of life fix for our on-call to see and investigate bot errors.

Fixes #3819