feat: allow onDemand configuration for release PRs (#5073)
This feature allows a repo to configure `onDemand: true` which will only
create the release PRs when you use the `release-please:force-run`
label. It will tag the releases on pushes to the release branch (e.g.
`main`).

This is useful for monorepos that choose when they want to run releases
(i.e. they do not always want a release PR as soon as a feature is
merged). Additionally, it should reduce the sheer number of API calls
the app makes for monorepos that have a set release schedule.

To configure, you can add `onDemand: true` to your
`.github/release-please.yml`. Example:

```
manifest: true
onDemand: true
branches:
- branch: 1.x
  manifest: true
  onDemand: true
```

**Note**: this option makes it so that your release PR could be "out of
date" with the latest code in the release branch, so maintainers will
need to pay close attention.