fix: don't open bug when .OwlBot.yaml is missing
The contents of the repo are the source of truth.  scan-googleapis-gen-and-create-pull requests queries firestore to find .OwlBot.yaml files.  Firestore may have stale data.  So if Owl Bot queries firestore, finds an .OwlBot.yaml, and then can't find the .OwlBot.yaml in the repo, then that's not actually a problem.

Fixes https://github.com/googleapis/repo-automation-bots/issues/5242