feat: bootstrapper CLI accepts --source-cl arg and passes as env var (#5287)
When triggering the owlbot-bootstrapper GCB job, require a `--source-cl`
argument. This is passed to the GCB job via a substitution which is
passed to the `pre-process` and `post-process` CLI steps as an argument
and the language container as an environment variable. The
`post-process` step uses the `--source-cl` argument and adds a
`PiperOrigin-RevId` to the generated PR body.