feat: add ability to comment on pull request (#5308)
Fixes #5307

Adds a new `commentInstructions` boolean configuration for the bot
(opt-in feature). If set and the PR author is not "trusted", comment on
the PR with reminder instructions for the maintainers on how to apply
labels/comment in order to trigger CI (whatever the bot would do for a
trusted contributor).