fix: bazelbot to fail if there are non-zero failing targets (#5561)
b/376777535 and b/377519778

There's a presubmit check in the source of the Bazel project. If someone
skips the check, then subsequent checks all become red and Bazel bot is
fine to fail in that case. We have to fix the root cause.

The test is also updated to this behavior.

---------

Co-authored-by: Anthonios Partheniou <partheniou@google.com>