deps: update gcf-utils oktokit dependencies to fix ReDoS vulnerability (#5669)
This is to fix ReDoS vulnerability found in octokit dependencies. This
was logic in choosing version:

Latest Probot version (13.4.3) is using @octokit/core 5.0.2 and
octokit/Plugin-paginate-rest: 9.2.2, which was a limiting factor in
upgrades:

@octokit/request: upgrade to 8.4.1 for vulnerability fix + compatibility
with octokit/core v5
@octokit/request-error: 5.1.1 for vulnerability fix
@octokit/plugin-paginate-rest: 9.2.2 for vulnerability fix 

Could not upgrade @octokit/rest > 19 because of incompatibility with
@octokit/plugin-paginate-rest > v9 , so need to manually set version of
@octokit/core dependency in package-lock.json to 5.0.2

Additional changes had to be made:

Webhook verify throws error if payload is not include, so added logic to
catch and propigate as a 500.

Tested via canary-bot:
https://github.com/googleapis/repo-automation-bots/issues/5672